Removed user and login
