+++ Source/Relativity.Transfer.SDK.Sample/App.config
@@ -7,11 +7,11 @@
 
 		<!-- URL of your RelativityOne environment i.e. https://contoso.relativity.one -->
 		<!-- Value shouldn't have any path, query, fragments or trailing slash - just schema with host. -->
-		<add key="RelativityOneInstanceUrl" value="" />
+		<add key="RelativityOneInstanceUrl" value="https://reg-prod.r1.kcura.com" />
 
 		<!-- UNC path to your RelativityOne fileshare *without* \Files\ suffix (can't have trailing slash. -->
 		<!-- This is taken from RelativityOne -> Servers (of 'Fileshare' type) tab. -->
-		<add key="RelativityOneFileshareRoot" value="" />
+		<add key="RelativityOneFileshareRoot" value="\\files.t008.ctus014128.r1.kcura.com\T008\Files" />
 
 		<!-- The location where the files are to be transferred relative to the root of the fileshare.
 			 You can modify this value and observe the transfer result in for example RelativityOne Staging Explorer).
@@ -20,10 +20,8 @@
 
 		<add key="DefaultSourceFilePath" value=""/>
 		<add key="DefaultSourceDirectoryPath" value=""/>
-
-		<add key="ClientLogin" value="" />
-		<add key="ClientOAuth2Id" value="" />
-		<add key="ClientSecret" value="" />
-		<add key="ClientPassword" value="" />
+		
+		<add key="ClientOAuth2Id" value="82ff9f193a134c08b268b14758396f0c" />
+		<add key="ClientSecret" value="adaec8773cd814220aac55139d887284fe8510d2" />
 	</appSettings>
 </configuration>
