+++ Source/Relativity.Transfer.SDK.Sample/Helpers/SettingNames.cs
@@ -9,8 +9,6 @@
 		public const string DefaultSourceFilePath = "DefaultSourceFilePath";
 		public const string DefaultSourceDirectoryPath = "DefaultSourceDirectoryPath";
 		public const string ClientSecret = "ClientSecret";
-		public const string ClientLogin = "ClientLogin";
-		public const string ClientPassword = "ClientPassword";
 		public const string ClientOAuth2Id = "ClientOAuth2Id";
 	}
 }
